Change block addresses to 64 bit numbers.

+++ src/block_device.rs
@@ -16,7 +16,7 @@ pub unsafe trait BlockDevice<const BLOCK_SIZE: usize> {
     /// # Safety
     ///
     /// `out`'s length must be at least a BLOCK_SIZE.
-    unsafe fn read_unchecked(&self, bnum: u32, out: &mut [u8]);
+    unsafe fn read_unchecked(&self, bnum: u64, out: &mut [u8]);
 
     /// Writes the data in `inp` to the block `bnum`.
     ///
@@ -25,15 +25,15 @@ pub unsafe trait BlockDevice<const BLOCK_SIZE: usize> {
     /// # Safety
     ///
     /// `inp`'s length must be at least a BLOCK_SIZE.
-    unsafe fn write_unchecked(&mut self, bnum: u32, inp: &[u8]);
+    unsafe fn write_unchecked(&mut self, bnum: u64, inp: &[u8]);
 
     /// Returns the number of blocks associated with this device.
-    fn num_blocks(&self) -> u32;
+    fn num_blocks(&self) -> u64;
 
     /// Reads the block `bnum` into `out`.
     ///
     /// This function may panic if `bnum` is out of bounds or if `out` is smaller than BLOCK_SIZE.
-    fn read_slice(&self, bnum: u32, out: &mut [u8]) {
+    fn read_slice(&self, bnum: u64, out: &mut [u8]) {
         assert!(
             out.len() >= BLOCK_SIZE,
             "Slice must be at least the length of the block"
@@ -46,7 +46,7 @@ pub unsafe trait BlockDevice<const BLOCK_SIZE: usize> {
     /// Writes the data in `inp` to the block `bnum`.
     ///
     /// This function may panic if `bnum` is out of bounds or if `inp` is smaller than BLOCK_SIZE.
-    fn write_slice(&mut self, bnum: u32, inp: &[u8]) {
+    fn write_slice(&mut self, bnum: u64, inp: &[u8]) {
         assert!(
             inp.len() >= BLOCK_SIZE,
             "Slice must be at least the length of the block"
@@ -59,7 +59,7 @@ pub unsafe trait BlockDevice<const BLOCK_SIZE: usize> {
     /// Reads the data in the block into the `out` by effectively performing a memcpy.
     ///
     /// This is safe because of the invariant for a type to implement `BlockData`.
-    fn read<T>(&self, bnum: u32, out: &mut T)
+    fn read<T>(&self, bnum: u64, out: &mut T)
     where
         T: BlockData<BLOCK_SIZE>,
     {
@@ -76,7 +76,7 @@ pub unsafe trait BlockDevice<const BLOCK_SIZE: usize> {
     /// Read data in the block and returns it as T.
     ///
     /// This is safe because of the invariant for a type to implement `BlockData`.
-    fn read_into<T>(&self, bnum: u32) -> T
+    fn read_into<T>(&self, bnum: u64) -> T
     where
         T: BlockData<BLOCK_SIZE>,
     {
@@ -96,7 +96,7 @@ pub unsafe trait BlockDevice<const BLOCK_SIZE: usize> {
     }
 
     /// Writes the data in in `inp` to disk by effectively performing a memcpy.
-    fn write<T>(&mut self, bnum: u32, inp: &T)
+    fn write<T>(&mut self, bnum: u64, inp: &T)
     where
         T: BlockData<BLOCK_SIZE>,
     {
@@ -144,7 +144,7 @@ mod tests {
 
     impl<const BLOCK_SIZE: usize> MemoryDevice<BLOCK_SIZE> {
         /// Constructs a new MemoryDevice.
-        pub fn new(num_blocks: u32) -> Self {
+        pub fn new(num_blocks: u64) -> Self {
             let num_blocks = num_blocks as usize;
             let capacity = num_blocks * BLOCK_SIZE;
             Self {
@@ -154,7 +154,7 @@ mod tests {
     }
 
     unsafe impl<const BLOCK_SIZE: usize> BlockDevice<BLOCK_SIZE> for MemoryDevice<BLOCK_SIZE> {
-        unsafe fn read_unchecked(&self, bnum: u32, out: &mut [u8]) {
+        unsafe fn read_unchecked(&self, bnum: u64, out: &mut [u8]) {
             assert!(bnum < self.num_blocks());
             let bnum = bnum as usize;
 
@@ -167,7 +167,7 @@ mod tests {
             }
         }
 
-        unsafe fn write_unchecked(&mut self, bnum: u32, inp: &[u8]) {
+        unsafe fn write_unchecked(&mut self, bnum: u64, inp: &[u8]) {
             assert!(bnum < self.num_blocks());
             let bnum = bnum as usize;
 
@@ -181,8 +181,8 @@ mod tests {
         }
 
         /// Returns the number of blocks associated with this memory device.
-        fn num_blocks(&self) -> u32 {
-            (self.buffer.len() / BLOCK_SIZE) as u32
+        fn num_blocks(&self) -> u64 {
+            (self.buffer.len() / BLOCK_SIZE) as u64
         }
     }
